[blog] Fix missing card
See how https://twitter.com/MUI_hq/status/1688919247504957440 doesn't have a card.
It's a follow-up on https://github.com/mui/material-ui/pull/38258.

I think that we should be able to fail the build if the card is missing, but too lazy.
