+++ docs/pages/experiments/blog/blog.md
@@ -3,7 +3,7 @@ title: Blog post title
 description: Our internationally distributed startup gathered on a remote island to get to know each other better. Here's what happened!
 date: 2022-07-28T00:00:00.000Z
 authors: ['samuelsycamore']
-card: true
+card: false
 tags: ['Company']
 ---
