+++ docs/public/static/blog/discord-announcement/discord.png
